+++ src/providers/INatPaperProvider.js
@@ -1,6 +1,6 @@
 // @flow
 
-import INatIcon from "components/INatIcon";
+import { INatIcon } from "components/SharedComponents";
 import type { Node } from "react";
 import * as React from "react";
 import { MD3LightTheme as DefaultTheme, Provider as PaperProvider } from "react-native-paper";
